Added functionality to login and registration activities. Working to solve FireBaseException.
