+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
         android:allowBackup="true"
@@ -14,8 +14,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyPantry"
         tools:targetApi="31">
+        <activity android:name=".MainActivity"/>
+        <activity android:name=".Profile.LoginActivity"/>
+        <activity android:name=".Profile.RegistrationActivity"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".splash_activity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
