+++ app/src/main/java/com/example/mypantry/splash_activity.java
@@ -1,0 +1,48 @@
+package com.example.mypantry;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.view.WindowManager;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import com.example.mypantry.Profile.LoginActivity;
+import com.google.firebase.FirebaseApp;
+
+public class splash_activity extends AppCompatActivity {
+
+    private static final int SPLASH = 2000;
+
+    Animation topAnim, bottomAnim;
+    ImageView imageView;
+    TextView textView;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
+        setContentView(R.layout.app_splash_activity);
+
+        topAnim = AnimationUtils.loadAnimation(this, R.anim.top_animation);
+        bottomAnim = AnimationUtils.loadAnimation(this, R.anim.bottom_animation);
+        imageView = findViewById(R.id.imageView);
+        textView = findViewById(R.id.text);
+
+        imageView.setAnimation(topAnim);
+        textView.setAnimation(bottomAnim);
+
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent = new Intent(splash_activity.this, LoginActivity.class);
+                startActivity(intent);
+                finish();
+            }
+        },SPLASH);
+    }
+}
