+++ build.gradle
@@ -18,3 +18,7 @@ plugins {
     id 'com.android.application' version '7.2.2' apply false
     id 'com.android.library' version '7.2.2' apply false
 }
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
+}
